Generate provider version constraint from Provider resource
Having to copy a provider file from somewhere would have been brittle
again. It's easier to carry the version constraint as part of the
provider resource and write it during synth. This has the benefit,
that only providers which are used in the stack are actually fetched.

In addition to this, it's introducing a `.terraform` in the project
root which gets linked into the `outdir` during synth. This allows
to remove outdir, without having to init again.