Fix pre release
With the upgrade of jsii to 1.14 a new version mapping for pre releases got introduced.
This broke our current schema based on git shas. This switches to GITHUB_RUN_NUMBER, which
is an incrementing number and stable for retries of the same workflow run (see https://docs.github.com/en/free-pro-team@latest/actions/reference/environment-variables)

With that change, we should be able to actually publish pre releases for Python as well.