Cleanup cli project templates
This drops `aws` as the pre defined provider, since we shouldn't
assume which provider the user wants to use. As we don't have
and providers set, we can omit `get` and `build` during init. This
will make the overall experience for users better and as a side effect
improves testing times.

Fixes #505