fix tests: add timeout to beforeAll hook to ensure jest waits for setup
apparently jest does not wait for an async function to complete in BeforeAll hooks (see: https://github.com/facebook/jest/issues/9527). With the default timeout of 5 seconds this had the effect of our test cases being run without initialization to be complete. this was no issue prior as we used to use execSync which blocked the whole js main loop and thereby also blocked jest in running our test cases.

switched exec to spawn for some test-helper shell calls. this allows to capture all logs and can also allow to print them as they appear (line per line) which helped discover the cause of the jest related issue