Don't require a scope to be present
Triggered by a discussion with @ansgarm how the Go feature would have
been scoped. We considered adding an explicit `global` scope, as well as
introducing a `jsii` or `languages` scope. However, we came to the
conclusion that this will become bloated very quickly. Hence, for
things which are touching essentially everything, we thought it's better
to just go without an explicit scope.