feat(cli): implement watch command using a GraphQL server in a seperately spawned process
The initial implementation of the new `cdktf watch` command also contains an
experiment in regards to separating the underlying invocation of a synth and of
Terraform for init, plan and apply from the presentation layer (which is in our
case a CLI build with React components using `ink`). The other commands (e.g.
`cdktf deploy`) which we currently have, make use of React Hooks for
orchestrating the underlying steps that are required. Quite early it turned out
to be unreasonable to implement the dynamic nature of the `watch` command using
those same tools. Instead we opted for trying out a separate process and using a
GraphQL api to connect the two. The main reason for this (as opposed to e.g.
ProtoBuf or REST) was the superior tooling support that exists for this in
TypeScript and React. However, with this prototype now in the open, we will
carefully examine it and assess the pros and cons before jumping ship.