fix(tests): use fs-extra for file copying to solve bug when running in Docker
The Asset construct failed to copy the local-asset.txt file while syntesizing after we upgraded to Node 14. It created the target file but then it blocked and the target file stayed empty. There were no visible permission errors on the source file local-asset.txt (it even worked with other files created by hand, which had the same permissions according to `ls -la` and `stat`) and it also worked on Windows and OS X. So this is going to stay a mystery for a now but we are happy that this works now