feat: support set for modules and improve set support for providers
for providers sets are exposed as lists (as they previously were) but when setting or retrieving them, the returned token will be wrapped in toset or tolist respectively. As sets are not yet supported by JSII, we will convert them to lists so that accessing individual elements works as inferred from the array type which is currently used to model sets in TypeScript. Addresses for example #424

fix(provider-generator): don't supply third arg for String|Number|Boolean Map types. remove return type to let TS narrow it down automatically to give a narrower type