feat(hcl2cdk): use Go based expression parser
This required us to shift entirely to async TS, but the implementation is now
directly coupled to HCL making it much more robust.
To limit the scope of this change I stayed with the search and replace notion of
the previous version, we will eventually need to shift this towards a rewrite model
when we want to support terraform functions through convert.

Closes #842