fix(cli): Don't validate types within node_modules
Related to https://github.com/yargs/yargs-parser/issues/452 which caused synth to fail. After this change TypeScript will only check types that are actually used in user code but not whether all types in dependencies are correct. We opted for this instead of pinning a version of yargs-parser without this bug because disabling this kind of checking will also help with future bugs that might occur and won't require an update of the TypeScript template. Thus this change also helps with #1991