fix(cli): handle missing package versions for non-npm packages
The previous implementation assumes that for every released npm version
there is an matching release on the other package managers. This sometimes
is not true, mostly by errors in the publishing process. To hide this and make
the UX of provider add nicer we will deploy the latest compatible matching version.

Closes #2065