fix(lib): Defer retrieving the Stack to a later time
As we moved away from `instanceof` we can't check for a subclass in the constructor of its own superclass. The Object.setProperty call will only be called after super finished in all base classes and this made it impossible for us to detect whether the instantiated construct was a Backend (which we use to display an additional to help users with a misconfiguration when creating a backend construct within the app and not within a stack). The stack.test.ts has a testcase for this behavior and caught this.

Co-authored-by: Daniel Schmidt <danielmschmidt92@gmail.com>