chore: trigger CI by labels only
A new action helps by assigning labels based on the changed files.
A new set of labels makes it easy to distinguish which unit tests should be run.
We also set the trigger to only one target (the one coresponding to label changes), so that
no CI run is double triggered and nothing needs to be aborted.