chore: increase operations per run of stale workflow
The `stale` action defaults to 30 `operations-per-run` to stay under GitHub's hourly rate limit, but for us, that's only enough to get us to issue 1697 in reverse cron before it runs out of operations and stops the job; therefore, older issues and PRs are not currently being checked by this workflow.

However, I looked it up and supposedly the rate limit for GitHub Enterprise is 15,000 requests per hour, so I think we can afford to go quite a bit higher. Let's try 200 for now, I'll monitor for another few days and then if we're not having issues maybe we can go up to ~1000 for future-proofing.