chore: fix finalizing the Sentry release
By piping to `/dev/null` we can get rid of this Sentry output:
`| (unreleased) | cdktf-cli-0.14.1        | 7          | 7 hours ago    |`
which was part of `status` previously (and required the newline replacement).

By grepping for `unreleased` and inverting the expression, we catch the
expected unreleased release which is the right condition to be able to
finalize the release (the previous one was wrong, as at this time, there
should always have been a release as it is created by an earlier workflow step).