feat: also go through files not yet added to git
previously the script would only use git ls-files in a way that listed all files known to git (to skip going through e.g. node_modules). However, this had the quirk, that files that hadn't been added to git yet, would not be detected which was annoying when writing documentation examples. This change fixes that.