feat(cli): support migrating state between backends
After this change a --migrate-state flag can be passed to diff, deploy, destroy
that will migrate the state from the previous backend to the current one.

Please not that Terraform CLI does not support migrating from the cloud configuration
to other backends, as a result this is not supported in cdktf either.

Closes #490
Closes #489