chore: fix CODEOWNERS
Heimdall has already been helpful to me in flagging there was an error with our CODEOWNERS file; someone had departed the company and no longer has write access to the repo. We should avoid naming individuals and try to use teams instead for that reason.

Also removed the line about branch protection rules because that may stop being true in the future as per new security policy that will be rolling out.