fix: correctly handle if version constraint is defined as star
Terraform does not support the star, instead it requires no version to be passed. This failed e.g. when initializing from a Terraform project as our version detection logic expects star to work. Fixing it like this has the benefit that users can use a star in the cdktf.json config as well (which currently would fail with the same error).