fix(lib): implement a central terraform module asset
This asset is at most once initiated per stack and copies all
local terraform modules in the cdktf.json in a single asset.

By doing this we can maintain the relative structure of the individual
modules.
A caveat is that we rely on information from the CLI (as we don't want to read
the cdktf.json in the library), but we can easily detect this and advise the
user to set the context or use the CLI.