fix(provider-generator): Skip statement block for rule block in aws_wafv2_web_acl and aws_wafv2_rule_group resources
This is a breaking change for users of these properties in these resources. It is necessary to bring down the size of the generated Go code though and also has an effect on all other languages.

For Go it shrinks the size of the generated code for the AWS provider from 648MB to 357MB. Previously the two resources accounted for 187MB (web acl) and 103MB (rule group) of the total size

Our resource docs already state that some attributes are skipped which includes mentions of the web acl resources (even though they weren't affected in the past)