fix: Make `CDKTF_LOG_FILE_DIRECTORY` a common flag available on all commands (#3495)
<!--

Unless this is a very simple 1-line-of-code change, please create a new
issue describing the change you're proposing first, then link to it from
this PR.

Read more about our process in our contributing guide:
https://github.com/hashicorp/terraform-cdk/blob/main/CONTRIBUTING.md

-->

### Related issue

Fixes #3489 

### Description

We were using `CDKTF_LOG_FILE_DIRECTORY` within our logger, but we never
made it into a supported CLI flag. CDKTF configures
[yargs](https://yargs.js.org/) in a way that all environment variables
with the `CDKTF_` prefix are treated as CLI flags. For commands that had
[`strict`](https://yargs.js.org/docs/#api-reference-strictenabledtrue)
enabled, that means that setting any non-supported flag with the
`CDKTF_` prefix would cause the command to fail. This PR now adds
support for this flag and it should work across all commands.

---------

Co-authored-by: Ansgar Mertens <ansgar@hashicorp.com>