fix: #3630 Due to skipping of attributes, we need to render certain blocks better
This is a side-effect of our effort to reduce provider size from blowing
up. We use a skip list to ignore a very small number of attributes that
are known to bloat the size of the provider significantly.

This doesn't cause a problem for our JSON synthesis, since it doesn't
care too much about types, but the HCL synthesis is more sensitive.

This commit, while not fixing the problem, will do two things. First,
it tries its best to output the attribute contents, and secondly,
lets the user know this is a problem and one that they might want to fix
either manually, or just use JSON output