chore(docs): remove documentation that has been moved to hashicorp/web-unified-docs
Back in May, all the contents of the website/ directory which powers our documentation on developer.hashicorp.com were migrated to the centralized hashicorp/web-unified-docs repository. At the time, we left the docs here just in case something happened and we needed to revert the migration.

Now that a couple of months have passed, we can confirm the migration was completed successfully and there's no reason to revert it, so it's time to completely delete all the contents of the old website/ directory, which can be a potential source of confusion for users.